backported integrationt test diff
